+++ arrus/core/devices/us4r/probeadapter/ProbeAdapterImplTest.cpp
@@ -116,6 +116,10 @@ public:
     MOCK_METHOD(void, disableAfeDemod, (), (override));
     MOCK_METHOD(uint16_t, getAfe, (uint8_t address), (override));
     MOCK_METHOD(void, setAfe, (uint8_t address, uint16_t value), (override));
+    MOCK_METHOD(float, getUCDMeasuredVoltage, (uint8_t), (override));
+    MOCK_METHOD(float, getFPGAWallclock, (), (override));
+    MOCK_METHOD(void, setHpfCornerFrequency, (uint32_t), (override));
+    MOCK_METHOD(void, disableHpf, (), (override));
 };
 
 class AbstractProbeAdapterImplTest : public ::testing::Test {